requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
python-decouple==3.6
requests==2.28.1
urllib3==1.26.13
